software/requirements.txt
fastapi==0.111.0
opencv_contrib_python==4.9.0.80
opencv_python==4.8.1.78
Pillow==10.1.0
Pillow==10.3.0
PyAudio==0.2.14
pydantic==2.7.4
Requests==2.32.3
urllib3==2.2.2
websocket_client==0.57.0
